handle search by health and filtration

+++ SoftExpertAssement/Model/SearchRecipeModel.swift
@@ -58,7 +58,7 @@ struct Recipe: Codable {
     let source: String?
     let url: String?
     let shareAs: String?
-    let yield: Int?
+    let yield: Double?
     let dietLabels: [String]?
     let healthLabels: [String]?
     let cautions: [String]?
